pyproject.toml
[project]
name = "voxtream-experiments"
version = "0.1.0"
description = "Stream-first runner bakeoff for Voxtream and Voxtream2 outside the main Aether Voice repo."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "fastapi>=0.115.12",
  "httpx>=0.28.1",
  "orjson>=3.10.18",
  "pydantic>=2.11.3",
  "python-dotenv>=1.1.0",
  "soundfile>=0.13.1",
  "typer>=0.16.1",
  "uvicorn>=0.34.0",
]

[project.scripts]
voxtream-experiments = "voxtream_experiments.cli:app"

[project.optional-dependencies]
native = [
  "voxtream>=0.2",
]

[dependency-groups]
dev = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voxtream_experiments"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv]
default-groups = ["dev"]
